Tablas maestras y creación de ausencias

+++ app/Http/Controllers/EpsController.php
@@ -6,11 +6,16 @@ use App\Models\Eps;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 use Inertia\Inertia;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class EpsController extends Controller
 {
+    use AuthorizesRequests;
+
     public function index(Request $request)
     {
+        $this->authorize('eps.view');
+
         $query = Eps::query()
             ->select('id', 'name', 'code', 'mobility_code', 'nit', 'regime')
             ->orderBy('name');
@@ -26,17 +31,25 @@ class EpsController extends Controller
 
         return Inertia::render('Eps/Index', [
             'eps' => $query->paginate(25)
-                          ->withQueryString()
+                          ->withQueryString(),
+            'can' => [
+                'create' => $request->user()->can('eps.create'),
+                'edit' => $request->user()->can('eps.edit'),
+                'delete' => $request->user()->can('eps.delete'),
+            ]
         ]);
     }
 
     public function create()
     {
+        $this->authorize('eps.create');
         return Inertia::render('Eps/Create');
     }
 
     public function store(Request $request)
     {
+        $this->authorize('eps.create');
+
         $validator = Validator::make($request->all(), [
             'name' => 'required|string|max:255|unique:eps',
             'code' => 'required|string|max:255|unique:eps',
@@ -55,8 +68,18 @@ class EpsController extends Controller
             ->with('success', 'EPS creada exitosamente.');
     }
 
+    public function show(Eps $ep)
+    {
+        $this->authorize('eps.view');
+        
+        return Inertia::render('Eps/Show', [
+            'eps' => $ep
+        ]);
+    }
+
     public function edit(Eps $ep)
     {
+        $this->authorize('eps.edit');
         return Inertia::render('Eps/Edit', [
             'eps' => $ep
         ]);
@@ -64,6 +87,8 @@ class EpsController extends Controller
 
     public function update(Request $request, Eps $ep)
     {
+        $this->authorize('eps.edit');
+
         $validator = Validator::make($request->all(), [
             'name' => 'required|string|max:255|unique:eps,name,' . $ep->id,
             'code' => 'required|string|max:255|unique:eps,code,' . $ep->id,
@@ -84,6 +109,8 @@ class EpsController extends Controller
 
     public function destroy(Eps $ep)
     {
+        $this->authorize('eps.delete');
+        
         $ep->delete();
 
         return redirect()->route('eps.index')
